bundling-scripts: consolidate logging
The logging calls currently carry along a lot of boilerplace, eg.
adding identions, output redirection, etc, which makes the code
hard to read and maintain.

Therefore encapsulate the actual logging in a small helper, which
is called this way:

    log 1 some text

The first parameter is the indention level (the number of tabs),
following ones are writen as-is.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>