bundling-scripts: helper for retrieving target installation dir
Adding a little function for retrieving the full path of an target
installation directory. This helps reducing the boilerplace and
not having the carry along the long prefix everywhere in the code.

When the function is called with a relative path name, it returns
that path with the prefix prepended, otherwiese just the prefix.

Example:

    cp foo $(target_dir opt/zimbra/conf)

This patch just introduces the helper function, following patches
will make use of it.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>