bundlings-scripts: fix missing package build of zimbra-timezones
The build process was yet not building a vital package, which
zimbra-core is depending on: zimbra-timezones.

Adding it to the package list and a script for building it.
The actual build/packaging is already done in zm-timezones
repo, so we just need to call it's build script and copy
over the generated .deb file.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>